requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
pyperclip